Reorganise unusued hooks for future
